rewrite: WIP rewrite to version 2.
