+++ src/models/api/statistics/article/dehydrated_reference.py
@@ -1,4 +1,4 @@
-from typing import List, Dict, Any
+from typing import Any, Dict, List
 
 from pydantic import BaseModel
