+++ src/models/api/statistics/reference/__init__.py
@@ -1,4 +1,4 @@
-from typing import List, Any, Dict
+from typing import Any, Dict, List
 
 from pydantic import BaseModel
