--- src/models/api/statistics/references-count/__init__.py
@@ -1,10 +1,0 @@
-from pydantic import BaseModel
-
-
-class ReferencesCountStatistics(BaseModel):
-    """This only returns the total number of references"""
-
-    total: int
-
-    class Config:
-        extra = "forbid"
