+++ src/models/api/statistics/references/references_job.py
@@ -1,0 +1,6 @@
+from src.models.api.job import Job
+
+
+class ReferencesJob(Job):
+    offset: int = 0
+    wari_id: str = ""
