+++ src/models/api/statistics/references/references_schema.py
@@ -1,0 +1,7 @@
+from marshmallow import Schema
+from marshmallow.fields import Int, String
+
+
+class ReferencesSchema(Schema):
+    offset = Int()
+    wari_id = String()
