+++ src/models/identifiers_checking/doi.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Optional, List, Dict, Any
+from typing import Any, Dict, List, Optional
 
 import pyalex
 from pyalex import Works
