+++ src/models/wikimedia/wikipedia/reference/person/english_wikipedia_template_person.py
@@ -1,7 +1,7 @@
-from src.models.wikimedia.wikipedia.reference.person import Person
 from src.models.wikimedia.wikipedia.reference.enums import (
     EnglishWikipediaTemplatePersonRole,
 )
+from src.models.wikimedia.wikipedia.reference.person import Person
 
 
 class EnglishWikipediaTemplatePerson(Person):
