--- src/views/statistics/references_count.py
@@ -1,6 +1,0 @@
-from src.views.statistics import StatisticsView
-
-
-class ReferencesCount(StatisticsView):
-    # TODO implement
-    raise NotImplementedError()
