+++ tests/test_doi.py
@@ -1,5 +1,7 @@
 from unittest import TestCase
 
+from src.models.identifiers_checking.doi import Doi
+
 
 class TestDoi(TestCase):
     retracted_in_oa_but_not_wd = "10.1038/nature04695"
