+++ tests/wikipedia/reference/test_person.py
@@ -1,10 +1,10 @@
 from unittest import TestCase
 
-from src.models.wikimedia.wikipedia.reference.person import Person
 from src.models.wikimedia.wikipedia.reference.enums import (
     EnglishWikipediaTemplatePersonRole,
     WikipediaTemplatePersonRole,
 )
+from src.models.wikimedia.wikipedia.reference.person import Person
 
 
 class TestPerson(TestCase):
